fix(api): move onInit to separate file
