+++ packages/api/.env.example
@@ -41,4 +41,4 @@ REPLY_TO=sg@gmail.com
 PASSWORD_RESET_TIME=600
 
 #Invitation
-INVITATION_EXPIRATION_DATE=10 days
+INVITATION_EXPIRATION_DATE="10 days"
